// functions/src/Models/DTO/Entrances.Drief.DTO.ts
export class EntrancesDriefDTO{

    proveedor_id: number;
    lotProveedor: string;
    product_id: number;
    date: string;
    quantity: string;
    observations: string;
    is_pz: boolean
    quality: boolean;
    rawMaterial: boolean;
    expiration: boolean;
    transport: boolean;
    strangeMaterial: boolean;
    odor: boolean;
    paking: boolean;
    color: boolean;
    texture: boolean;
    weigth: boolean;
}